[RFC] Move fast pass functions into a centralized file.
[ghstack-poisoned]
