+++ aten/src/ATen/native/Fill.cpp
@@ -8,54 +8,6 @@
 namespace at {
 namespace native {
 
-// ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ fill ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-namespace {
-  template <typename scalar_t>
-  inline void fill_fast(Tensor& self, Scalar value_scalar) {
-    auto value = value_scalar.to<scalar_t>();
-    scalar_t * dptr = static_cast<scalar_t *>(self.data_ptr());
-    *dptr = value;
-  }
-} // namspace
-
-Tensor& fill_out(Tensor& self, Scalar value) {
-  if (self.is_quantized()) {
-    at::Tensor out = at::ones(self.sizes()).to(kFloat) * value;
-    out = out.to(self.device());
-    // Trust the `copy_` to handle the quantization and the boundary chacks.
-    self.copy_(out);
-    return self;
-  }
-  // When filling a number to 1-element CPU tensor, we want to skip
-  // everything but manipulate data ptr directly.
-  // Ideally this fast pass should be implemented in TensorIterator,
-  // but we also want to skip compute_types which in not avoidable
-  // in TensorIterator for now.
-  if (self.device() == at::kCPU && self.numel() == 1 && !self.is_complex() && !value.isComplex()) {
-     AT_DISPATCH_ALL_TYPES_AND3(kHalf, kBool, kBFloat16, self.scalar_type(), "fill_out", [&]() {
-        fill_fast<scalar_t>(self, value);});
-     return self;
-  }
-  auto iter = TensorIteratorConfig()
-    .set_check_mem_overlap(false)  // Fill is idempotent, so overlap is okay
-    .check_all_same_dtype(false)
-    .add_output(self)
-    .resize_outputs(false)
-    .build();
-  fill_stub(iter.device_type(), iter, value);
-  return self;
-}
-
-Tensor& fill_(Tensor& self, Scalar value) {
-  return fill_out(self, value);
-}
-
-Tensor& fill_(Tensor& self, const Tensor& value) {
-  TORCH_CHECK(value.dim() == 0, "fill_ only supports 0-dimension value tensor but got tensor with ", value.dim(), " dimensions.");
-  return fill_out(self, value.item());
-}
-
-DEFINE_DISPATCH(fill_stub);
 
 // ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ fill_diagonal ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
